Finish implementing Chord-1478 (power dialog) - still needs uinputto be made writable. (dm)

+++ b2g/input/src/org/nbp/b2g/input/Action.java
@@ -24,7 +24,7 @@ public abstract class Action {
     return ScreenMonitor.getScreenMonitor();
   }
 
-  protected boolean delay (int interval) {
+  protected boolean delay (long interval) {
     try {
       Thread.sleep(interval);
     } catch (InterruptedException exception) {
